Refactor and improve model and improve controller
Change how models work. Add a Collection trait for fungible/unidentifiable types. Collection is a supertrait of Model. Make all trait types and type constraints of trait types as loose as possible, so that you don't have to implement a trait for a type if it isn't used.

Fix errors so that they use 422 Unproccessable Entity for failed validation instead of 400 Bad Request.

Refactor macro and more.