Fixed a bug that prevented the db macro from working
It works now, but everything has been overhauled in a much more functional style instead of imperative. The two concerns of the main processing of the data and the codegen have been separated out, making it much easier to work with and understand. It also makes it much more flexible and decouples the two concerns, which previously made it very hard to add or change features, such as foreign keys, including mc<->mc relationships.