Add token auth to controller
Integrate auth logic to the controller. A custom type can be used to validate the parsed token. It is a blanket rule for all routes in a controller, it isn't currently possible to have different rules for different routes.

A column can be designated as unique. This will affect requests using this column as an index.